+++ src/python_version.rs
@@ -9,7 +9,7 @@ use std::path::Path;
 pub(crate) const DEFAULT_PYTHON_VERSION: PythonVersion = PythonVersion {
     major: 3,
     minor: 12,
-    patch: 4,
+    patch: 5,
 };
 
 /// Representation of a specific Python `X.Y.Z` version.